requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
selenium==4.17.2
schedule==1.2.0
pydantic==2.5.3
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1